chore(testing): switch from ava to jest
